// frontend/projects/vacme-web-generated/src/lib/model/ort-der-impfung-display-name-extended-jax.ts
/**
 * Generated VacMe API
 * Generated using custom templates to be found under vacme-web-generated/src/templates.
 *
 * The version of the OpenAPI document: 999.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrtDerImpfungTypTS } from './ort-der-impfung-typ';
import { LatLngJaxTS } from './lat-lng-jax';


export interface OrtDerImpfungDisplayNameExtendedJaxTS { 
    version?: number;
    timestampErstellt?: Date;
    timestampMutiert?: Date;
    userErstellt?: string;
    userMutiert?: string;
    id?: string;
    name?: string;
    terminverwaltung?: boolean;
    mobilerOrtDerImpfung?: boolean;
    booster?: boolean;
    deaktiviert?: boolean;
    oeffentlich?: boolean;
    glnNummer?: string;
    typ?: OrtDerImpfungTypTS;
    noFreieTermine1?: boolean;
    noFreieTermine2?: boolean;
    noFreieTermineN?: boolean;
    nextTermin1Date?: Date;
    nextTermin2Date?: Date;
    nextTerminNDate?: Date;
    latLng?: LatLngJaxTS;
    distanceToReg?: number;
    impfungGegenBezahlung?: boolean;
}